fix some logic

+++ cmd/api/coffeeHandlers.go
@@ -12,9 +12,7 @@ import (
 	"github.com/julienschmidt/httprouter"
 )
 
-// coffees returns a list of coffees to the user. The function
-// is optimized to store the coffee list retrieved from db
-// in the session data - thereby reducing db calls.
+// coffees returns a list of coffees to the user.
 func (app *application) coffees(w http.ResponseWriter, r *http.Request) {
 	// First check the session data to see if we already have the product list
 	// Note, the type assertion here will return a fully nil object
@@ -22,45 +20,10 @@ func (app *application) coffees(w http.ResponseWriter, r *http.Request) {
 	// Hence a little code duplication below.
 	coffeeList, ok := app.sessionManager.Get(r.Context(), "coffeeList").([]dataModels.Coffee)
 	if !ok {
-		// If coffeeList doesn't exists yet, we call the database.
-		// Define a timeout for the database retrieval
-		ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-		defer cancel()
-
-		// Call the database
-		coffeeList, err := dataModels.CoffeeList(ctx, app.db)
-		if err != nil {
-			app.logger.Error("dataModels.CoffeList", err)
-			http.Error(w, "Server error", http.StatusInternalServerError)
-			return
-		}
-
-		// Cache the coffee list in the session token
-		// this is to avoid subsequent database queries
-		// if list becomes too large - this is a bad optimization.
-		app.sessionManager.Put(r.Context(), "coffeeList", coffeeList)
-		coffees := make([]dataModels.Coffee, len(coffeeList))
-		for i, v := range coffeeList {
-			coffees[i] = dataModels.Coffee{
-				ID:          v.ID,
-				Name:        v.Name,
-				Description: v.Description,
-				Price:       v.Price,
-			}
-		}
-		// Format response to json
-		js, err := json.Marshal(coffees)
-		if err != nil {
-			app.logger.Error("marshal json", err)
-			http.Error(w, "Server error", http.StatusInternalServerError)
-			return
-		}
-		// Write response to http.ResponseWriter
-		js = append(js, '\n')
-		w.Header().Set("Content-Type", "application/json")
-		w.Write(js)
+		app.logger.Error("middleware failed to load coffee products")
+		http.Error(w, "Server error", http.StatusInternalServerError)
 		return
-	} // End initial DB Call for list of products
+	}
 
 	coffees := make([]dataModels.Coffee, len(coffeeList))
 	for i, v := range coffeeList {
@@ -87,8 +50,6 @@ func (app *application) coffees(w http.ResponseWriter, r *http.Request) {
 
 // coffeeDetails will return the specified coffee product
 // with additional details such as caffeine and calories.
-// This function attempts to avoid unnecessary db calls
-// by first checking the session for list of products.
 func (app *application) coffeeDetails(w http.ResponseWriter, r *http.Request) {
 	// Get the parameters from the request url context, ie ":id"
 	params := httprouter.ParamsFromContext(r.Context())
@@ -104,42 +65,10 @@ func (app *application) coffeeDetails(w http.ResponseWriter, r *http.Request) {
 	// If coffee list has not been pulled, need to pull it.
 	coffeeList, ok := app.sessionManager.Get(r.Context(), "coffeeList").([]dataModels.Coffee)
 	if !ok {
-		// app.logger.Error("Session doesn't contain coffeeList")
-		// http.Error(w, "Coffee products not available", http.StatusBadRequest)
-		// return
-		// If coffeeList doesn't exists yet, we call the database.
-		// Define a timeout for the database retrieval
-		ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-		defer cancel()
-
-		// Call the database
-		coffeeList, err := dataModels.CoffeeList(ctx, app.db)
-		if err != nil {
-			app.logger.Error("dataModels.CoffeList", err)
-			http.Error(w, "Server error", http.StatusInternalServerError)
-			return
-		}
-		app.sessionManager.Put(r.Context(), "coffeeList", coffeeList)
-		coffeeDesc := dataModels.Coffee{}
-		for _, coffee := range coffeeList {
-			if coffee.ID == id {
-				coffeeDesc = coffee
-				break
-			}
-		}
-
-		js, err := json.Marshal(coffeeDesc)
-		if err != nil {
-			app.logger.Error("marshal json", err)
-			http.Error(w, "Server error", http.StatusInternalServerError)
-			return
-		}
-		// Write response to http.ResponseWriter
-		js = append(js, '\n')
-		w.Header().Set("Content-Type", "application/json")
-		w.Write(js)
+		app.logger.Error("middleware failed to load coffee products")
+		http.Error(w, "Server error", http.StatusInternalServerError)
 		return
-	} // End initial DB Call for List of Products
+	}
 
 	coffeeDesc := dataModels.Coffee{}
 	for _, coffee := range coffeeList {
@@ -161,11 +90,9 @@ func (app *application) coffeeDetails(w http.ResponseWriter, r *http.Request) {
 	w.Write(js)
 }
 
+// newCoffee is an administrative function to allow
+// an authorized user to add a coffee product to the data base
 func (app *application) newCoffee(w http.ResponseWriter, r *http.Request) {
-	// We expect a content-type of application/json for this request
-	// However the determining factor will be whether the data can
-	// marshall to json struct or not.
-
 	var newCoffee dataModels.Coffee
 	err := json.NewDecoder(r.Body).Decode(&newCoffee)
 	if err != nil {
@@ -196,6 +123,10 @@ func (app *application) newCoffee(w http.ResponseWriter, r *http.Request) {
 	js = append(js, '\n')
 	w.Write(js)
 }
+
+// updateCoffee is an administrative function to
+// allow an authorized user to update an existing
+// coffee product in the database
 func (app *application) updateCoffee(w http.ResponseWriter, r *http.Request) {
 	// Get the parameters from the request url context, ie ":id"
 	params := httprouter.ParamsFromContext(r.Context())
@@ -215,7 +146,21 @@ func (app *application) updateCoffee(w http.ResponseWriter, r *http.Request) {
 
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
-
+	if ok := dataModels.CheckCoffeeExists(ctx, app.db, id); !ok {
+		notFound := fmt.Sprintf("coffee %d was not found", id)
+		msg := map[string]string{"Bad request": notFound}
+		js, err := json.Marshal(msg)
+		if err != nil {
+			app.logger.Error("marshal json", err)
+			http.Error(w, "Server error", http.StatusInternalServerError)
+			return
+		}
+		// Write response to http.ResponseWriter
+		js = append(js, '\n')
+		w.Header().Set("Content-Type", "application/json")
+		w.Write(js)
+		return
+	}
 	// Call the database
 	err = dataModels.UpdateCoffee(ctx, app.db, id, coffee)
 	if err != nil {
@@ -236,6 +181,10 @@ func (app *application) updateCoffee(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	w.Write(js)
 }
+
+// deleteCoffee is an administrative function to
+// allow an authorized user to delete a coffee product
+// from the database.
 func (app *application) deleteCoffee(w http.ResponseWriter, r *http.Request) {
 	// Get the parameters from the request url context, ie ":id"
 	params := httprouter.ParamsFromContext(r.Context())
@@ -247,7 +196,21 @@ func (app *application) deleteCoffee(w http.ResponseWriter, r *http.Request) {
 	}
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
-
+	if ok := dataModels.CheckCoffeeExists(ctx, app.db, id); !ok {
+		notFound := fmt.Sprintf("coffee %d was not found", id)
+		msg := map[string]string{"Bad request": notFound}
+		js, err := json.Marshal(msg)
+		if err != nil {
+			app.logger.Error("marshal json", err)
+			http.Error(w, "Server error", http.StatusInternalServerError)
+			return
+		}
+		// Write response to http.ResponseWriter
+		js = append(js, '\n')
+		w.Header().Set("Content-Type", "application/json")
+		w.Write(js)
+		return
+	}
 	// Call the database
 	err = dataModels.DeleteCoffee(ctx, app.db, id)
 	if err != nil {
